arrangement File - Home 페이지 컴포넌트 식으로 section 분리 - Layout은 Notion참조
