+++ src/components/Home/PhoneIcon.jsx
@@ -1,0 +1,16 @@
+import React from 'react'
+import { PhoneFilled } from '@ant-design/icons';
+import styled from 'styled-components'
+
+function PhoneIcon() {
+    const PhoneStyle = styled(PhoneFilled)`
+        transform: rotateY(180deg); 
+    `
+    return (
+        <>
+            <PhoneStyle/>
+        </>
+    )
+}
+
+export default PhoneIcon
