--- src/components/Ul.jsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-const Ul = styled.ul`
-    list-style:none;
-    margin: 0;
-    padding: 0;
-`
-
-export default Ul;
